refactor: reduce public exports in response to review
- Remove defaultReadFile from command-registry.ts exports (internal detail)
- Remove re-export passthrough in input.ts (parseFrontmatter, resolveContent,
  resolveCommandFilePath, listSkillNames, ListDir, CommandSuggestion no longer
  tunnelled through input.ts — callers import from command-registry.ts directly)
- Remove resolveCommandFilePath, resolveContent, ListDir from index.ts exports
  (they are now implementation details of CommandRegistry)
- Update tests to import from command-registry.ts instead of input.ts
